Fix the three CI failures
Staging ANGLE from build.rs was a race. Cargo does not order a crate's
own build script against its dependencies' build scripts, so mozangle's
OUT_DIR did not exist yet on a clean CI build; locally it always did.
The build scripts now pick the DLLs out of that OUT_DIR after cargo
finishes, and build.rs is gone.

scripts/build.sh had no executable bit in the index — chmod +x on Windows
does not reach git without core.fileMode.

The lint job built all of Servo in the dev profile and the build job built
it again in release, so a cold run cost over two hours. Windows now runs
one debug pass through build.ps1 -Checks, and formatting moved to a Linux
job that needs no dependency build. Release builds stay in the release
workflow.

-Checks now passes -D warnings, so it can act as the gate CI relies on.

+++ src/webview_node.rs
@@ -722,7 +722,7 @@ fn godot_key_to_servo(event: &Gd<InputEventKey>) -> Option<Key> {
         _ => {
             // 印字可能な文字はそのまま渡す。
             let unicode = event.get_unicode();
-            let character = char::from_u32(unicode as u32).filter(|c| !c.is_control())?;
+            let character = char::from_u32(unicode).filter(|c| !c.is_control())?;
             return Some(Key::Character(character.to_string()));
         }
     };
